Added a new dataset that has 2 fields for partitionColumn The HIVE_TABLE_OPT_KEY is by default the name of the dataset Hive details are currently defined in the DataSetDef - It should be moved from here. There are 2 classes that deal with extracting the partition path - one is a dumb SlashToDash partition extractor can be used for single field - the other (SlashSplits) is required for multifield partitions I removed the spaces in the names of the partition - I got errors on the HDFS so - no spaces in the HDFS paths
Nota bene:
This commit expects the files on the HDFS
Use the SparkBuilder and the DataSetDef.commonOpts to configure you environment.
This version uses the 0.4.3-y hudi version which is found here:  https://github.com/arw357/hudi/tree/docker-yields
This version should be changed to 0.4.4 when the new Uber version gets here.